requirements.txt
aiomysql==0.2.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-jazzmin==3.0.1
django-modeltranslation==0.19.12
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient==2.2.7
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
